+++ docs/changes.md
@@ -29,11 +29,11 @@ Changes in command line arguments:
 Release date: `2018-06-xx`
 
 ### Core
+- [NXDRIVE-605](https://jira.nuxeo.com/browse/NXDRIVE-605): Handle corrupted SQLite database
 - [NXDRIVE-1230](https://jira.nuxeo.com/browse/NXDRIVE-1230): Fix the ability to disable a Worker
 
 ### Packaging / Build
 - [NXDRIVE-1229](https://jira.nuxeo.com/browse/NXDRIVE-1229): Do not install Drive in %APPDATA% but %LOCALAPPDATA% folder on Windows
-
 
 
 # 3.1.0